finishing connected droppable tests
